refactor: merge types.rs into traits.rs, add unimock to PcmReader
Consolidate AudioEffect and PcmReader traits into a single traits module.
Add unimock mock generation for PcmReader (PcmReaderMock) and export it
from the mock module. Add explicit lifetime to read_planar for unimock
compatibility.
