+++ crates/kithara-play/src/impls/player_resource.rs
@@ -264,7 +264,7 @@ mod tests {
             to_write
         }
 
-        fn read_planar(&mut self, output: &mut [&mut [f32]]) -> usize {
+        fn read_planar<'a>(&mut self, output: &'a mut [&'a mut [f32]]) -> usize {
             if self.eof || output.is_empty() {
                 return 0;
             }
