+++ crates/kithara-play/src/impls/player_track.rs
@@ -394,7 +394,7 @@ mod tests {
             buf.len()
         }
 
-        fn read_planar(&mut self, output: &mut [&mut [f32]]) -> usize {
+        fn read_planar<'a>(&mut self, output: &'a mut [&'a mut [f32]]) -> usize {
             if output.is_empty() {
                 return 0;
             }
