+++ crates/kithara-play/src/impls/resource.rs
@@ -177,7 +177,7 @@ impl Resource {
     }
 
     /// Read deinterleaved (planar) PCM samples.
-    pub fn read_planar(&mut self, output: &mut [&mut [f32]]) -> usize {
+    pub fn read_planar<'a>(&mut self, output: &'a mut [&'a mut [f32]]) -> usize {
         self.inner.read_planar(output)
     }
 
@@ -330,7 +330,7 @@ mod tests {
             to_write
         }
 
-        fn read_planar(&mut self, output: &mut [&mut [f32]]) -> usize {
+        fn read_planar<'a>(&mut self, output: &'a mut [&'a mut [f32]]) -> usize {
             if self.eof || output.is_empty() {
                 return 0;
             }
